insert and view and edit product validation

+++ src/main/java/ir/alzahra/offerBaz/control/impl/OfferServiceImpl.java
@@ -3,12 +3,9 @@ package ir.alzahra.offerBaz.control.impl;
 
 import ir.alzahra.offerBaz.control.IOfferCheckService;
 import ir.alzahra.offerBaz.control.IOfferService;
-import ir.alzahra.offerBaz.dto.BankDTO;
-import ir.alzahra.offerBaz.dto.searchParameter.ProductSearchParam;
 import ir.alzahra.offerBaz.exception.BaseException;
 import ir.alzahra.offerBaz.model.dao.IBankDao;
 import ir.alzahra.offerBaz.model.dao.IProductDao;
-import ir.alzahra.offerBaz.model.dao.impl.ProductDao;
 import ir.alzahra.offerBaz.model.entity.BankEntity;
 import ir.alzahra.offerBaz.model.entity.ProductEntity;
 import ir.alzahra.offerBaz.notify.CustomSpringEvent;
@@ -42,12 +39,6 @@ public class OfferServiceImpl implements IOfferService {
 
     private String finalcode;
 
-    @Override
-    public void insertProduct(ProductEntity productEntity) throws BaseException{
-        productDao.insert(productEntity);
-        applicationEventPublisher.notify("product.insert.success", NotificationType.Info);
-
-    }
 
     public String generateUniqueCode(String name2) throws BaseException {
        long code = bankDao.getCountOfRecord()+1000;
@@ -77,7 +68,10 @@ public class OfferServiceImpl implements IOfferService {
 
     @Override
     public List<BankEntity> getAllBanks() throws BaseException {
-       return bankDao.getAllBanks();
+        List<BankEntity> banks=bankDao.getAllBanks();
+        if (Objects.isNull(banks)|| banks.size()==0)
+            throw new BaseException("bank.find.emptyBankList");
+       return banks;
     }
 
     @Override
@@ -98,20 +92,32 @@ public class OfferServiceImpl implements IOfferService {
 
     @Override
     public ProductEntity findProductByCode(String trackingCode) throws BaseException {
-        return productDao.findProductByCode(trackingCode);
+        ProductEntity p=productDao.findProductByCode(trackingCode);
+        if (Objects.isNull(p))
+            throw new BaseException("product.insert.notFoundProduct");
+        else return p;
     }
 
     @Override
     public void updateProduct(ProductEntity productEntity) throws BaseException {
+        if (Objects.isNull(productEntity.getId()))
+            throw new BaseException("product.edit.notFoundForEdit");
+        else{
+            offerCheckService.checkUpdateProduct(productEntity);
+            checkExistProduct(productEntity.getProductName());
         productDao.update(productEntity);
         applicationEventPublisher.notify("product.update.success", NotificationType.Info);
+        }
     }
 
     @Override
     public void deleteProduct(ProductEntity productEntity) throws BaseException {
-        productDao.delete(productEntity);
-        applicationEventPublisher.notify("product.delete.success", NotificationType.Info);
-
+        if (Objects.isNull(productEntity.getId()))
+            throw new BaseException("product.delete.notFoundForDelete");
+        else {
+            productDao.delete(productEntity);
+            applicationEventPublisher.notify("product.delete.success", NotificationType.Info);
+        }
     }
 
     @Override
@@ -131,5 +137,13 @@ public class OfferServiceImpl implements IOfferService {
         if(Objects.isNull(bankEntity) || bankEntity.size()==0)
             throw new BaseException("bank.find.notFoundBank");
         else return bankEntity;
+    }
+
+    @Override
+    public void checkExistProduct(String name) throws BaseException {
+        ProductEntity p =productDao.findByName(name);
+        if (Objects.nonNull(p))
+            throw new BaseException("product.insert.exist");
+
     }
 }
